Add the RouterContainer class
DHC-1517

The problem in DHC-1517 seems to be a race condition between
the thread that is deleting the state-machine and the healtcheck
thread. In order to remove this race condtions, a first step is to
create a container class that contains data structures for state
machines storage and manages insert and delete actions with a lock.

Change-Id: I5972a1cd7290761a985810cdfdb56dc1c49144b5
Signed-off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>